Fix 4 Linux compat bugs found during audit (with regression tests)
Bug 1 (active): VirtualProtect(size=0) mprotected a full page instead of
being a no-op. Windows docs: 'If the size is 0, the function returns
non-zero and does not modify the page protection.' Added early return.

Bug 2 (active): map_view_of_pe had no bounds check on the PE section
table. A malformed PE with large SizeOfOptionalHeader or NumberOfSections
caused a heap buffer over-read (segfault). Added validation before the
section iteration loop.

Bug 4 (active): VirtualQuery on MEM_FREE regions reported Protect=PAGE_NOACCESS
(0x01) instead of 0 per Windows spec.

Bug 5 (active): VirtualQuery on MEM_FREE regions returned the raw query
address as BaseAddress instead of page-aligning it.

Bug 3 (latent): _BitScanForward/_BitScanReverse passed unsigned long
(64-bit on LP64) to __builtin_ctz/__builtin_clz (documented for
unsigned int). For mask > UINT_MAX, implicit narrowing to 0 caused UB
(__builtin_clz(0)). Fixed by explicit (uint32_t) cast. No callers in
kananlib, so this was latent.

Bug 6 (unproven): /proc/self/maps fgets 512-byte buffer could drop
regions with long paths. Changed to getline(). Cannot demonstrate failure
(controlled by kernel), kept as defensive improvement.

Test: 19 regression tests in TestLinuxAudit.cpp. 4 bugs demonstrated with
revert->segfault/fail->re-apply->pass methodology. 14/14 full suite pass.

+++ include/compat/intrin.h
@@ -18,14 +18,17 @@
 // translation units that include both <intrin.h> and <windows.h>.
 
 // Bit scans (MSVC returns 0/1 and writes the index through a pointer).
+// On LP64 Linux, unsigned long is 64-bit but MSVC's is 32-bit. Cast to
+// uint32_t to preserve MSVC semantics and avoid UB from __builtin_clz(0)
+// when the caller passes a value wider than 32 bits.
 static inline unsigned char _BitScanForward(unsigned long* index, unsigned long mask) {
     if (mask == 0) return 0;
-    *index = (unsigned long)__builtin_ctz(mask);
+    *index = (unsigned long)__builtin_ctz((uint32_t)mask);
     return 1;
 }
 static inline unsigned char _BitScanReverse(unsigned long* index, unsigned long mask) {
     if (mask == 0) return 0;
-    *index = (unsigned long)(31 - __builtin_clz(mask));
+    *index = (unsigned long)(31 - __builtin_clz((uint32_t)mask));
     return 1;
 }
 static inline unsigned char _BitScanForward64(unsigned long* index, unsigned long long mask) {
